fix(feed): address CodeRabbit review feedback on PR #13
6 issues raised by coderabbitai:

[Critical] FeedItem.is_ai NOT NULL backfill
- Use @Column(columnDefinition = "BOOLEAN NOT NULL DEFAULT FALSE")
  so existing rows are auto-populated when ddl-auto=update adds the column.

[Major] FeedDetailResponse parent field redeclaration
- Removed all 22 redeclared parent fields. With @SuperBuilder, the inherited
  builder already exposes setters. Kept only the 6 detail-specific fields
  (plan, priceBreakdown, preparation, venueAnchors, primaryPin,
  confirmedPartnerProfiles).
- Rewrote from(...) to build FeedDetailResponse directly via the inherited
  builder API, populating parent fields from FeedItem (no longer constructing
  an intermediate FeedItemResponse).

[Major] FeedItem URL columns default to VARCHAR(255)
- Added @Column(length = 2048) on imageUrl and authorAvatarUrl.

[Minor] FeedItemResponse.calculateProgressPercent int overflow
- Cast to long before multiplying: (int) ((long) fundedAmount * 100L / fundingGoal)

[Minor] FeedParticipantProfile/FeedAuthorProfile snake_case missing
- Added @JsonProperty("avatar_url") on avatarUrl in both DTOs.

[Minor] PriceBreakdown.baseFee primitive int loses null semantics
- Changed to Integer to match AddOn.price and preserve null distinction.

+++ capstone-api/src/main/java/backend/feed/dto/FeedItemResponse.java
@@ -137,6 +137,6 @@ public class FeedItemResponse {
 			return null;
 		}
 		Integer fundedAmount = feedItem.getFundedAmount() == null ? 0 : feedItem.getFundedAmount();
-		return fundedAmount * 100 / fundingGoal;
+		return (int) ((long) fundedAmount * 100L / fundingGoal);
 	}
 }
